requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
Faker>=18.0.0
django-filter>=23.0
